# docs/gateway-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "summary": "Health check",
        "tags": ["Health"]
      }
    },
    "/api/users": {
      "post": {
        "description": "Rota **pública** (sem token). Cria conta + carteira, gera senha/CodigoCliente/ChaveLoja e envia o acesso por e-mail. A API responde apenas com mensagem de sucesso.",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateUserDto" } }
          }
        },
        "responses": { "201": { "description": "" } },
        "summary": "Cadastro público de usuário (PF/PJ)",
        "tags": ["Users"]
      }
    },
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Dados do usuário autenticado",
        "tags": ["Users"]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Rota **pública**. Use o documento e a senha recebidos por e-mail no cadastro. Retorna Bearer token + CodigoCliente + ChaveLoja para as demais rotas.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginDto" } }
          }
        },
        "responses": { "201": { "description": "" } },
        "summary": "Login público (CPF/CNPJ + senha)",
        "tags": ["Auth"]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "description": "Rota **pública**. Gera nova senha e envia no e-mail cadastrado (documento + e-mail devem coincidir).",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ResetPasswordDto" } }
          }
        },
        "responses": { "201": { "description": "" } },
        "summary": "Reset de senha (público)",
        "tags": ["Auth"]
      }
    },
    "/api/wallet": {
      "get": {
        "operationId": "WalletController_getWallet",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Consultar saldo da carteira",
        "tags": ["Wallet"]
      }
    },
    "/api/wallet/transactions": {
      "get": {
        "description": "Filtros opcionais: status (APPROVED, DENIED, PENDING, EXPIRED, CANCELLED) e type (PIX, CREDIT_CARD, WITHDRAWAL).",
        "operationId": "WalletController_listTransactions",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "example": 50, "type": "string" }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["PENDING", "APPROVED", "DENIED", "EXPIRED", "CANCELLED"],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": { "enum": ["PIX", "CREDIT_CARD", "WITHDRAWAL"], "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Extrato de transações da carteira",
        "tags": ["Wallet"]
      }
    },
    "/api/payments/pix": {
      "post": {
        "description": "Gera EMV (copia e cola) + qrCodeBase64. Aceita externalReference para conciliar a venda. Dispara webhook PAYMENT_PIX se cadastrado. Status APPROVED/DENIED aleatório.",
        "operationId": "PaymentsController_createPix",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreatePixPaymentDto" } }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Pagamento Pix / QR Code (simulado)",
        "tags": ["Payments"]
      }
    },
    "/api/payments/card": {
      "post": {
        "description": "Obrigatório consultar GET /api/fees e enviar `installments` (1–21) + `feePercent` da tabela. Bandeiras: Visa, Mastercard, Elo. Valor líquido (bruto − taxa) credita a carteira se APPROVED. Dispara webhook PAYMENT_CARD.",
        "operationId": "PaymentsController_createCard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCardPaymentDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Pagamento cartão de crédito (simulado)",
        "tags": ["Payments"]
      }
    },
    "/api/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Consultar pagamento por ID",
        "tags": ["Payments"]
      }
    },
    "/api/webhooks": {
      "post": {
        "description": "Um URL por evento (PAYMENT_PIX, PAYMENT_CARD, WITHDRAWAL). Após cada transação, a Lera Box envia POST JSON com status APPROVED/DENIED.",
        "operationId": "WebhooksController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateWebhookDto" } }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Cadastrar / atualizar webhook",
        "tags": ["Webhooks"]
      },
      "get": {
        "operationId": "WebhooksController_list",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Listar webhooks cadastrados",
        "tags": ["Webhooks"]
      }
    },
    "/api/webhooks/{id}": {
      "delete": {
        "operationId": "WebhooksController_remove",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Remover webhook",
        "tags": ["Webhooks"]
      }
    },
    "/api/fees": {
      "get": {
        "description": "Tabela fixa Visa / Mastercard / Elo — parcelas 1 a 21. Use estes valores no pagamento com cartão (`installments` + taxa correspondente).",
        "operationId": "FeesController_list",
        "parameters": [
          {
            "name": "brand",
            "required": false,
            "in": "query",
            "description": "Filtrar por bandeira",
            "schema": { "enum": ["VISA", "MASTERCARD", "ELO"], "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "summary": "Consultar taxas de cartão (público)",
        "tags": ["Fees"]
      }
    },
    "/api/withdrawals": {
      "post": {
        "description": "Se o saldo real for insuficiente, nega com INSUFFICIENT_BALANCE. Caso contrário, simula aprovação/negação. Aprovação debita a carteira.",
        "operationId": "WithdrawalsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateWithdrawDto" } }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Solicitar saque (simulado)",
        "tags": ["Withdrawals"]
      }
    },
    "/api/withdrawals/{id}": {
      "get": {
        "operationId": "WithdrawalsController_findOne",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Consultar saque por ID",
        "tags": ["Withdrawals"]
      }
    }
  },
  "info": {
    "title": "Lera Box",
    "description": "Ambiente de simulação para testes VBA Systems. Este ambiente monitorado para integração, simulando um Gateway de pagamento, tem como mostrar ao profissional um pouco da estrutura que trabalhamos dentro da nossa empresa.\n\n**Acesso:** `POST /api/users` é público. Após o cadastro, o profissional recebe senha, CodigoCliente e ChaveLoja por e-mail e usa `POST /api/auth/login` (CPF/CNPJ + senha) para obter o Bearer token das demais rotas.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "personType": { "type": "string", "enum": ["PF", "PJ"], "example": "PF" },
          "name": {
            "type": "string",
            "example": "Maria Silva",
            "description": "Nome completo (PF) ou razão social (PJ)"
          },
          "tradingName": {
            "type": "string",
            "example": "Loja da Maria",
            "description": "Nome fantasia (recomendado para PJ)"
          },
          "email": {
            "type": "string",
            "example": "maria@empresa.com",
            "description": "E-mail real — recebe senha e notificações"
          },
          "phone": {
            "type": "string",
            "example": "11999998888",
            "description": "Celular brasileiro válido com DDD (11 dígitos). CPF/CNPJ podem ser fictícios; telefone e e-mail devem ser reais."
          },
          "document": {
            "type": "string",
            "example": "12345678901",
            "description": "CPF (11) ou CNPJ (14) — pode ser fictício (sandbox). Apenas formato."
          },
          "zipCode": {
            "type": "string",
            "example": "01310100",
            "description": "CEP apenas dígitos"
          },
          "address": { "type": "string", "example": "Av. Paulista" },
          "number": { "type": "string", "example": "1000" },
          "complement": { "type": "string", "example": "Sala 12" },
          "neighborhood": { "type": "string", "example": "Bela Vista" },
          "city": { "type": "string", "example": "São Paulo" },
          "state": { "type": "string", "example": "SP" }
        },
        "required": [
          "personType",
          "name",
          "email",
          "phone",
          "document",
          "zipCode",
          "address",
          "number",
          "neighborhood",
          "city",
          "state"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "document": {
            "type": "string",
            "example": "12345678901",
            "description": "CPF (11) ou CNPJ (14), apenas dígitos"
          },
          "password": { "type": "string", "example": "A1b2C3d4!" }
        },
        "required": ["document", "password"]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "document": {
            "type": "string",
            "example": "12345678901",
            "description": "CPF ou CNPJ cadastrado"
          },
          "email": {
            "type": "string",
            "example": "maria@empresa.com",
            "description": "E-mail cadastrado — recebe a nova senha"
          }
        },
        "required": ["document", "email"]
      },
      "CreatePixPaymentDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 15000,
            "description": "Valor em centavos (ex: 15000 = R$ 150,00)"
          },
          "description": { "type": "string", "example": "Pagamento pedido #123" },
          "payerDocument": {
            "type": "string",
            "example": "12345678901",
            "description": "CPF/CNPJ do pagador (pode ser fictício)"
          },
          "externalReference": {
            "type": "string",
            "example": "PEDIDO-123",
            "description": "Referência externa da venda (seu ID de pedido). Usada para conciliar webhook e consulta."
          }
        },
        "required": ["amount", "payerDocument"]
      },
      "CreateCardPaymentDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 25000,
            "description": "Valor bruto em centavos (ex: 25000 = R$ 250,00)"
          },
          "description": { "type": "string", "example": "Compra loja online" },
          "externalReference": {
            "type": "string",
            "example": "PEDIDO-456",
            "description": "Referência externa da venda para conciliar no webhook"
          },
          "cardNumber": {
            "type": "string",
            "example": "4111111111111111",
            "description": "Número do cartão (Visa, Mastercard ou Elo)"
          },
          "cardHolder": { "type": "string", "example": "MARIA SILVA" },
          "expiryMonth": { "type": "string", "example": "12" },
          "expiryYear": { "type": "string", "example": "2030" },
          "cvv": { "type": "string", "example": "123" },
          "installments": {
            "type": "number",
            "example": 3,
            "minimum": 1,
            "maximum": 21,
            "description": "Parcelas (1–21). Deve existir na tabela GET /api/fees"
          },
          "feePercent": {
            "type": "number",
            "example": 3.19,
            "description": "Taxa percentual consultada em GET /api/fees para a bandeira + parcelas. Deve coincidir exatamente com a taxa cadastrada, senão a transação é rejeitada."
          }
        },
        "required": [
          "amount",
          "cardNumber",
          "cardHolder",
          "expiryMonth",
          "expiryYear",
          "cvv",
          "installments",
          "feePercent"
        ]
      },
      "CreateWebhookDto": {
        "type": "object",
        "properties": {
          "event": {
            "type": "string",
            "enum": ["PAYMENT_PIX", "PAYMENT_CARD", "WITHDRAWAL"],
            "example": "PAYMENT_PIX",
            "description": "Um endpoint por evento (PIX, cartão ou saque)"
          },
          "url": {
            "type": "string",
            "example": "https://seu-sistema.com/webhooks/lera-box/pix",
            "description": "URL HTTPS que receberá o POST do webhook"
          },
          "secret": {
            "type": "string",
            "example": "segredo-opcional-para-hmac",
            "description": "Se informado, enviamos header X-Lera-Box-Signature = HMAC-SHA256(body, secret)"
          }
        },
        "required": ["event", "url"]
      },
      "CreateWithdrawDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 10000,
            "description": "Valor do saque em centavos"
          },
          "pixKey": {
            "type": "string",
            "example": "00020126580014br.gov.bcb.pix...",
            "description": "Chave Pix de destino (e-mail, CPF, telefone ou EVP)"
          },
          "description": { "type": "string", "example": "Saque para conta pessoal" },
          "externalReference": {
            "type": "string",
            "example": "SAQUE-001",
            "description": "Referência externa para conciliar no webhook"
          },
          "document": {
            "type": "string",
            "example": "12345678901",
            "description": "CPF do titular da chave Pix"
          }
        },
        "required": ["amount", "pixKey", "document"]
      }
    }
  }
}